gv_device: don't compare guint32 and guint16.
packet count is a 16 bit value in gvcp packets.

+++ src/arvgvdevice.c
@@ -42,7 +42,7 @@ static GRegex *arv_gv_device_url_regex = NULL;
 typedef struct {
 	GMutex *mutex;
 
-	guint32 packet_count;
+	guint16 packet_count;
 
 	GSocket *socket;
 	GSocketAddress	*interface_address;
